feat: block ambiente delete when some paciente exists

+++ src/controllers/ambiente/delete-ambiente/delete-ambiente.ts
@@ -1,6 +1,9 @@
 import { AmbienteNotFoundError } from "../../../errors/ambiente-not-found-error.js";
+import { AmbienteWithPacientesError } from "../../../errors/ambiente-with-pacientes-error.js";
 import { MongoDeleteAmbienteRepository } from "../../../repositories/ambiente/delete-ambiente/mongo-delete-ambiente.js";
 import { MongoGetAmbienteRepository } from "../../../repositories/ambiente/get-ambiente/mongo-get-ambiente.js";
+import { MongoListPacienteRepository } from "../../../repositories/paciente/list-paciente/mongo-list-paciente.js";
+import type { IListPacienteRepository } from "../../paciente/list-paciente/types.js";
 import { JwtTokenController } from "../../token/jwt-token-controller.js";
 import type { IGetAmbienteRepository } from "../get-ambiente/types.js";
 import type { IDeleteAmbienteController, IDeleteAmbienteRepository } from "./types.js";
@@ -9,6 +12,7 @@ export class DeleteAmbienteController implements IDeleteAmbienteController {
   constructor(
     private readonly getAmbienteRepository: IGetAmbienteRepository = new MongoGetAmbienteRepository(),
     private readonly deleteAmbienteRepository: IDeleteAmbienteRepository = new MongoDeleteAmbienteRepository(),
+    private readonly listPacienteRepository: IListPacienteRepository = new MongoListPacienteRepository(),
   ) {}
 
   async handle(idAmbiente: string, authHeader?: string): Promise<void> {
@@ -20,7 +24,18 @@ export class DeleteAmbienteController implements IDeleteAmbienteController {
       throw new AmbienteNotFoundError();
     }
 
-    // TODO: validar se não tem pacientes no ambiente
+    const pacientes = await this.listPacienteRepository.listPaciente(
+      {
+        offset: 0,
+        limit: 1,
+      },
+      idAplicacao,
+    );
+    const hasPacientesOnThisAmbiente = pacientes.length === 1;
+    if (hasPacientesOnThisAmbiente) {
+      throw new AmbienteWithPacientesError();
+    }
+
     await this.deleteAmbienteRepository.deleteAmbiente(ambiente.id, idAplicacao);
   }
 }
